+++ common/src/main/java/io/github/xienaoban/biologydictionary/core/skill/general/GetSpawnEggSkill.java
@@ -5,6 +5,7 @@ import io.github.xienaoban.biologydictionary.Lang;
 import io.github.xienaoban.biologydictionary.core.skill.GeneralSkill;
 import io.github.xienaoban.biologydictionary.core.skill.SkillCost;
 import io.github.xienaoban.biologydictionary.platform.util.EntityUtils;
+import io.github.xienaoban.biologydictionary.platform.util.ItemUtils;
 import io.github.xienaoban.biologydictionary.platform.util.PlayerUtils;
 import io.github.xienaoban.biologydictionary.platform.util.TextUtils;
 import net.fabricmc.api.EnvType;
@@ -44,7 +45,7 @@ public record GetSpawnEggSkill(EntityType<?> entityType) implements GeneralSkill
         if (entityType == null) {
             BiologyDictionary.sendCenteredWarning(ctx.player(), TextUtils.translate(Lang.TEXT_UNKNOWN_ENTITY_TYPE));
         } else {
-            SpawnEggItem item = SpawnEggItem.byId(entityType);
+            SpawnEggItem item = ItemUtils.getSpawnEggItem(entityType);
             if (item == null) {
                 BiologyDictionary.sendCenteredWarning(ctx.player(), TextUtils.translate(Lang.TEXT_NO_DATA_WITH_BRACKETS));
             } else {